requirements.txt
selenium
pyperclip
pyyaml
